Modifying files for migration to boshv2 schema
