+++ config/templates/blueprint-manifest.yml.ejs
@@ -1,4 +1,4 @@
-<%= spec.header.select({disk_pools: [p('disk_pool')], resource_pools: [{name: p('resource_pool')}]}) %>
+<%= spec.header %>
 
 update:
   canaries: 0
@@ -7,23 +7,11 @@ update:
   update_watch_time: 1000-100000
   serial: false
 
-jobs: &jobs
+instance_groups: &instance_groups
 <%
 const net = spec.networks[0];
+const stemcell = spec.stemcell;
 const members = net.static.slice(0,1);
-%>
-- name: blueprint_<%= net.az %>
-  instances: <%= members.length %>
-  networks:
-  - name: <%= net.name %>
-    static_ips: <%= JSON.stringify(members) %>
-  resource_pool: <%= `${p('resource_pool')}_${net.az}` %>
-  persistent_disk_pool: <%= p('disk_pool') %>
-  templates:
-  - { name: blueprint, release: blueprint }
-  - { name: blueprint-agent, release: blueprint }
-
-<%
 const properties = {};
 if (!spec.previous_manifest) {
   /* create scenario: pick random service instance properties */
@@ -34,21 +22,35 @@ if (!spec.previous_manifest) {
     }
   };
 } else {
-  /* update scenario: preserve previous service instance properties */
-  properties.blueprint = spec.previous_manifest.properties.blueprint;
+  if(!spec.previous_manifest.instance_groups){
+    properties.blueprint = spec.previous_manifest.properties.blueprint;
+  }else{
+    properties.blueprint = spec.previous_manifest.instance_groups[0].jobs[0].properties;
+  }
 }
 %>
-properties:
+- name: blueprint
+migrated_from: 
+- name: blueprint_z1
+instances: <%= members.length %>
+networks:
+- name: <%= net.name %>
+  static_ips: <%= JSON.stringify(members) %>
+vm_type: <%= p('vm_type') %>
+persistent_disk_type: <%= p('disk_type') %>
+stemcell: <%= stemcell.alias %>
+jobs:
+- name: blueprint
+  release: blueprint
   blueprint: &blueprint
+  properties:
     port: 8080
     admin:
       username: <%= properties.blueprint.admin.username %>
       password: <%= properties.blueprint.admin.password %>
-  agent:
+- name: blueprint-agent
+  release: blueprint
+  properties:
     username: <%= p('agent.auth.username') %>
     password: <%= p('agent.auth.password') %>
     provider: <%= JSON.stringify(p('agent.provider')) %>
-    manifest:
-      jobs: *jobs
-      properties:
-        blueprint: *blueprint
