+++ lib/bosh/manifest/Network.js
@@ -7,6 +7,7 @@ const _ = require('lodash');
 class Network {
   constructor(opts) {
     this.name = opts.name;
+    this.subnet_name = opts.subnet_name;
     this.type = opts.type;
     this.cloud_properties = opts.cloud_properties;
   }
