+++ lib/bosh/manifest/Networks.js
@@ -22,14 +22,15 @@ class Networks {
         const networkSegmentation = network.segmentation || defaultSegmentationOptions;
         const nets = network.subnets.map(subnet => {
           const net = Network.create(_.assign({
-            name: `${network.name}_${subnet.az}`,
+            name: network.name,
+            subnet_name: `${network.name}_${subnet.az}`,
             type: network.type,
             index: index,
             offset: networkSegmentation.offset || defaultSegmentationOptions.offset || DEFAULT_OFFSET,
             size: networkSegmentation.size || defaultSegmentationOptions.size || DEFAULT_SIZE,
           }, subnet));
           this.all.push(net);
-          this[net.name] = net;
+          this[net.subnet_name] = net;
           return net;
         });
         this[network.name] = nets;
