+++ lib/controllers/ServiceFabrikAdminController.js
@@ -227,7 +227,7 @@ class ServiceFabrikAdminController extends BaseController {
             this.director.getTasks({
               deployment: deploymentName
             }),
-            manager.diffManifest(deploymentName, opts).then(unifyDiffResult)
+            manager.diffManifest(deploymentName, opts).then(utils.unifyDiffResult)
           ])
           .spread((vms, tasks, diff) => ({
             name: deploymentName,
@@ -253,7 +253,7 @@ class ServiceFabrikAdminController extends BaseController {
       return _
         .chain(deployment)
         .pick('name', 'releases', 'stemcell')
-        .set('diff', unifyDiffResult(deployment))
+        .set('diff', utils.unifyDiffResult(deployment))
         .set('instance', _
           .chain(deployment.entity)
           .pick('name', 'last_operation', 'space_guid', 'service_plan_id', 'service_plan_guid', 'dashboard_url')
@@ -717,23 +717,6 @@ class ServiceFabrikAdminController extends BaseController {
 module.exports = ServiceFabrikAdminController;
 
 /* Helper functions */
-function unifyDiffResult(result) {
-  const diff = [];
-  _.each(result.diff, _.spread((value, type) => {
-    switch (type) {
-    case 'added':
-      diff.push(`+${value}`);
-      break;
-    case 'removed':
-      diff.push(`-${value}`);
-      break;
-    default:
-      diff.push(` ${value}`);
-      break;
-    }
-  }));
-  return diff;
-}
 
 function getPlanByGuid(plans, guid) {
   const planId = _
