+++ lib/controllers/ServiceFabrikApiController.js
@@ -22,6 +22,7 @@ const ScheduleManager = require('../jobs');
 const config = require('../config');
 const CONST = require('../constants');
 const catalog = require('../models').catalog;
+const utils = require('../utils');
 
 const CloudControllerError = {
   NotAuthorized: err => {
@@ -535,10 +536,11 @@ class ServiceFabrikApiController extends BaseController {
             .findDeploymentNameByInstanceId(req.params.instance_id)
             .then(deploymentName => this.cloudController.getOrgAndSpaceGuid(req.params.instance_id)
               .then(opts => req.manager.diffManifest(deploymentName, opts))
+              .then(result => utils.unifyDiffResult(result))
             )
             .then(result => {
-              scheduleInfo.update_required = result.diff && result.diff.length > 0;
-              scheduleInfo.update_details = result.diff;
+              scheduleInfo.update_required = result && result.diff.length > 0;
+              scheduleInfo.update_details = result;
               return scheduleInfo;
             });
         } else {
