+++ lib/fabrik/DirectorManager.js
@@ -352,6 +352,7 @@ class DirectorManager extends BaseManager {
       name: deploymentName,
       director_uuid: this.director.uuid,
       releases: this.releases,
+      stemcells: [this.stemcell],
       compilation: BoshDirectorClient.getInfrastructure().compilation,
       disk_pools: BoshDirectorClient.getInfrastructure().disk_types,
       resource_pools: this.resourcePools,
@@ -367,7 +368,8 @@ class DirectorManager extends BaseManager {
       properties: this.settings.context || {},
       previous_manifest: opts.previous_manifest,
       organization_guid: opts.organization_guid,
-      space_guid: opts.space_guid
+      space_guid: opts.space_guid,
+      stemcell: this.stemcell
     });
     if (networks[this.networkName] === undefined) {
       logger.error(`subnet ${this.networkName} definition not found among the applicable networks defintion : ${JSON.stringify(networks)}`);
