+++ lib/jobs/ServiceInstanceUpdateJob.js
@@ -75,10 +75,10 @@ class ServiceInstanceUpdateJob extends BaseJob {
     logger.info(`Instance Id: ${instanceDetails.instance_id} - manager : ${plan.manager.name} - Force Update: ${plan.service.force_update}`);
     return this
       .getOutdatedDiff(instanceDetails.deployment_name, instanceDetails.instance_id, plan)
-      .then(diffList => {
-        const outdated = diffList.length !== 0;
+      .then(diffResults => {
+        const outdated = _.get(diffResults, 'diff', false) && diffResults.diff.length !== 0;
         operationResponse.deployment_outdated = outdated;
-        operationResponse.diff = diffList;
+        operationResponse.diff = utils.unifyDiffResult(diffResults);
         if (!outdated) {
           operationResponse.update_init = 'NA';
           logger.info(`Instance: ${instanceDetails.instance_name} - Deployment: ${instanceDetails.deployment_name} up-to date. No update required.`);
@@ -88,7 +88,7 @@ class ServiceInstanceUpdateJob extends BaseJob {
         return this
           .updateDeployment(
             instanceDetails.deployment_name,
-            diffList,
+            diffResults.diff,
             plan.service.force_update)
           .then(result => {
             operationResponse.update_init = CONST.OPERATION.SUCCEEDED;
@@ -123,8 +123,7 @@ class ServiceInstanceUpdateJob extends BaseJob {
       .createManager(plan)
       .then((directorManager) => cloudController.getOrgAndSpaceGuid(instanceId)
         .then(opts => directorManager
-          .diffManifest(deploymentName, opts)
-          .then(result => result.diff))
+          .diffManifest(deploymentName, opts))
       );
   }
