+++ lib/utils/index.js
@@ -46,6 +46,7 @@ exports.isServiceFabrikOperation = isServiceFabrikOperation;
 exports.isServiceFabrikOperationFinished = isServiceFabrikOperationFinished;
 exports.taskIdRegExp = taskIdRegExp;
 exports.hasChangesInForbiddenSections = hasChangesInForbiddenSections;
+exports.unifyDiffResult = unifyDiffResult;
 
 function streamToPromise(stream, options) {
   const encoding = _.get(options, 'encoding', 'utf8');
@@ -352,4 +353,22 @@ function hasChangesInForbiddenSections(diff) {
     throw new errors.Forbidden(`Automatic update not possible. Detected changes in forbidden section(s) '${forbiddenSections.join(',')}'`);
   }
   return false;
+}
+
+function unifyDiffResult(result) {
+  const diff = [];
+  _.each(result.diff, _.spread((value, type) => {
+    switch (type) {
+    case 'added':
+      diff.push(`+${value}`);
+      break;
+    case 'removed':
+      diff.push(`-${value}`);
+      break;
+    default:
+      diff.push(` ${value}`);
+      break;
+    }
+  }));
+  return diff;
 }
